requirements.txt
wheel==0.36.2
Cython==0.29.21
mahotas==1.4.11
matplotlib==3.3.3
numpy==1.19.2
scipy==1.5.2
scikit-image==0.18.0
scikit-learn==0.24.0
opencv-python==4.4.0.46
certifi==2020.12.5
imageio==2.9.0
click==7.1.2
pandas==1.1.5
PyYAML==5.3.1
pytest==6.2.1